requirements.txt
flask
python-dotenv
logging